+++ src/lightning/pytorch/utilities/testing/_runif.py
@@ -24,7 +24,7 @@ from lightning.fabric.accelerators.cuda import num_cuda_devices
 from lightning.fabric.utilities.imports import _TORCH_GREATER_EQUAL_2_0, _TORCH_GREATER_EQUAL_2_1
 from lightning.pytorch.accelerators.cpu import _PSUTIL_AVAILABLE
 from lightning.pytorch.accelerators.mps import MPSAccelerator
-from lightning.pytorch.accelerators.tpu import TPUAccelerator
+from lightning.pytorch.accelerators.xla import XLAAccelerator
 from lightning.pytorch.callbacks.progress.rich_progress import _RICH_AVAILABLE
 from lightning.pytorch.core.module import _ONNX_AVAILABLE
 from lightning.pytorch.strategies.deepspeed import _DEEPSPEED_AVAILABLE
@@ -127,7 +127,7 @@ def _RunIf(
         reasons.append("unimplemented on Windows")
 
     if tpu:
-        conditions.append(not TPUAccelerator.is_available())
+        conditions.append(not XLAAccelerator.is_available())
         reasons.append("TPU")
         # used in conftest.py::pytest_collection_modifyitems
         kwargs["tpu"] = True